% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{down_trans_redcap}
\alias{down_trans_redcap}
\title{Downloads REDCap's data and transforms it}
\usage{
down_trans_redcap(token, uri = "https://biredcap.mh.org.au/api/")
}
\arguments{
\item{token}{Redcap api token}

\item{uri}{Redcap uri}
}
\description{
Downloads REDCap's data and transforms it
}
