% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_redcap_data}
\alias{get_redcap_data}
\title{Access the raw RedCap data table}
\usage{
get_redcap_data(token, uri = "https://biredcap.mh.org.au/api/")
}
\arguments{
\item{token}{Redcap api token}

\item{uri}{Redcap uri}
}
\description{
Will subset by access_group when it is not 'all'
}
