% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbl_participant}
\alias{get_tbl_participant}
\title{Extracts the participant table from the baseline table}
\usage{
get_tbl_participant(raw)
}
\arguments{
\item{raw}{Raw redcap data}
}
\description{
Everyone who has consented is a participant. I'm relying on the 'consent'
variable for this. The participant table contains every attribute that each
participant has one of except the screening-related attributes (not everyone
who is screened consents).
}
