% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbl_screening}
\alias{get_tbl_screening}
\title{Extracts the screening table from the baseline table}
\usage{
get_tbl_screening(baseline)
}
\arguments{
\item{baseline}{The baseline table}
}
\description{
Extracts the screening table from the baseline table
}
