% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbl_swab}
\alias{get_tbl_swab}
\title{Extracts the swab table}
\usage{
get_tbl_swab(raw_consented)
}
\arguments{
\item{raw_consented}{Consented subset of the REDCap table}
}
\description{
Extracts the swab table
}
