% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbl_symptom}
\alias{get_tbl_symptom}
\title{Extracts the symptom table from the raw list}
\usage{
get_tbl_symptom(raw_list)
}
\arguments{
\item{raw_list}{Raw data list}
}
\description{
Binds rows of all symptom surveys
}
