% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbl_withdrawal}
\alias{get_tbl_withdrawal}
\title{Extracts the withdrawal table}
\usage{
get_tbl_withdrawal(raw)
}
\arguments{
\item{raw}{Raw redcap data}
}
\description{
Extracts the withdrawal table
}
