% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{get_tbls}
\alias{get_tbls}
\title{Extracts all proper tables from the event tables}
\usage{
get_tbls(raw_list)
}
\arguments{
\item{raw_list}{List returned by \code{\link{raw_to_list}}}
}
\description{
Extracts all proper tables from the event tables
}
