% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_altnames.R
\docType{data}
\name{medicalhx_altnames}
\alias{medicalhx_altnames}
\title{Names for options in the medical history}
\format{
An object of class \code{character} of length 9.
}
\usage{
medicalhx_altnames
}
\description{
Names for options in the medical history
}
\keyword{datasets}
