% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{redcap_subset}
\alias{redcap_subset}
\title{Subsets raw REDCap table by access group defined by sites}
\usage{
redcap_subset(raw, access_group)
}
\arguments{
\item{raw}{Raw redcap data}

\item{access_group}{Access group to subset the data}
}
\description{
If all - doesn't do anything
If none - removes all rows
}
