% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(token, ...)
}
\arguments{
\item{token}{String or path to file containing the RedCap API token.}

\item{...}{Not used.}
}
\description{
Run the Shiny Application
}
