% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_altnames.R
\docType{data}
\name{site_altnames}
\alias{site_altnames}
\title{Access group names}
\format{
An object of class \code{character} of length 7.
}
\usage{
site_altnames
}
\description{
Access group names
}
\keyword{datasets}
