% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dataccess.R
\name{subset_consent}
\alias{subset_consent}
\title{Subsets the REDCap table to only those who consented}
\usage{
subset_consent(raw)
}
\arguments{
\item{raw}{Raw redcap data}
}
\description{
Subsets the REDCap table to only those who consented
}
