% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_altnames.R
\docType{data}
\name{swabres_altnames}
\alias{swabres_altnames}
\title{Names for options in swab results}
\format{
An object of class \code{character} of length 14.
}
\usage{
swabres_altnames
}
\description{
Names for options in swab results
}
\keyword{datasets}
