% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_altnames.R
\docType{data}
\name{var_altnames}
\alias{var_altnames}
\title{Alternative variable names}
\format{An object of class \code{character} of length 9.}
\usage{
var_altnames
}
\description{
Alternative variable names
}
\keyword{datasets}
