% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(token = "token", key = "key", ...)
}
\arguments{
\item{token}{Path to file containing the token}

\item{key}{Path to file containing the key used to create hashes}

\item{...}{Not used.}
}
\description{
Run the Shiny Application
}
